+++ src/main/java/ua/edu/sumdu/j2se/Nikolai/tasks/view/AddEditTaskView.java
@@ -5,7 +5,7 @@ import ua.edu.sumdu.j2se.Nikolai.tasks.controller.ControllerConstants;
 public class AddEditTaskView implements View {
 
     @Override
-    public long printInfo() {
+    public int printInfo() {
         System.out.println("Task added or edited");
         return ControllerConstants.MainControllerAction;
     }
