+++ src/main/java/ua/edu/sumdu/j2se/Nikolai/tasks/view/View.java
@@ -1,5 +1,5 @@
 package ua.edu.sumdu.j2se.Nikolai.tasks.view;
 
 public interface View {
-    long printInfo();
+    int printInfo();
 }
